Move derives into same nested crate
Will be getting more derives in the futures so makes sense to have them
all live together.

Setup inspired by [Diesel](https://github.com/diesel-rs/diesel/blob/master/diesel_derives/src/lib.rs)
