--- crates/enum-each-variant-derive/.gitignore
@@ -1,3 +1,0 @@
-
-/target
-**/*.rs.bk
