--- crates/enum-each-variant-derive/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "enum-each-variant-derive"
-version = "0.1.2"
-authors = ["David Pedersen <david.pdrsn@gmail.com>"]
-repository = "https://github.com/davidpdrsn/enum-each-variant-derive.git"
-homepage = "https://github.com/davidpdrsn/enum-each-variant-derive"
-license = "MIT"
-readme = "README.md"
-description = "Background jobs for Rust"
-
-[dependencies]
-syn = "0.12.14"
-quote = "0.4.2"
-
-[lib]
-proc-macro = true
