+++ src/lib.rs
@@ -1,11 +1,9 @@
 #![deny(missing_copy_implementations, trivial_casts, trivial_numeric_casts, unsafe_code,
         unstable_features, unused_import_braces, unused_qualifications)]
 
-#[macro_use]
-extern crate derive_enqueueable;
-#[macro_use]
-extern crate enum_each_variant_derive;
 extern crate redis;
+#[macro_use]
+extern crate robin_derives;
 extern crate serde;
 #[macro_use]
 extern crate serde_derive;
@@ -24,5 +22,5 @@ pub mod prelude {
     pub use connection::{establish, ConnectionProducer, WorkerConnection};
     pub use worker::boot;
     pub use config::Config;
-    pub use derive_enqueueable::*;
+    pub use robin_derives::*;
 }
